Catch AmazonS3Exception in BlobStoreManager upload naming and cr changes
