+++ src/main/scala/com.socrata.snapshotter/BlobManager.scala
@@ -4,7 +4,7 @@ import java.io.InputStream
 
 import com.amazonaws.event.{ProgressListener, ProgressEvent}
 import com.amazonaws.services.s3.AmazonS3Client
-import com.amazonaws.services.s3.model.{ObjectMetadata, PutObjectRequest}
+import com.amazonaws.services.s3.model.{AmazonS3Exception, ObjectMetadata, PutObjectRequest}
 import com.amazonaws.services.s3.transfer.TransferManager
 import com.amazonaws.services.s3.transfer.model.UploadResult
 
@@ -13,7 +13,7 @@ import com.rojoma.json.v3.ast.JValue
 
 import org.slf4j.LoggerFactory
 
-object BlobManager {
+object BlobStoreManager {
   private lazy val s3client = new AmazonS3Client()
   private lazy val manager = new TransferManager(s3client)
   private val logger = LoggerFactory.getLogger(getClass)
@@ -23,19 +23,29 @@ object BlobManager {
     manager.shutdownNow()
   }
 
-  def upload(inStream: InputStream, keyName: String): UploadResult = {
-    val req = new PutObjectRequest(SnapshotterConfig.awsBucketName, s"$keyName", inStream, new ObjectMetadata())
-    logger.info(s"Sending put request to s3: $req")
-    val upload = manager.upload(req)
+  def upload(inStream: InputStream, path: String): Either[JValue, UploadResult] = {
+    try {
+      val req = new PutObjectRequest(SnapshotterConfig.awsBucketName, s"$path", inStream, new ObjectMetadata())
+      logger.info(s"Sending put request to s3: $req")
+      val upload = manager.upload(req)
 
-    upload.addProgressListener(new SpecialListener(logger))
+      upload.addProgressListener(new LoggingListener(logger))
 
-    val uploadResult = upload.waitForUploadResult()
-    logger.info(s"uploadResult: $uploadResult")
-    uploadResult
+      val uploadResult = upload.waitForUploadResult()
+      logger.info(s"uploadResult: $uploadResult")
+      Right(uploadResult)
+    } catch {
+      case exception: AmazonS3Exception => Left(
+        json"""{ message: "Problem uploading to S3",
+                     error: ${exception.toString},
+                     "error code": ${exception.getErrorCode},
+                     "error type": ${exception.getErrorType},
+                     "error message": ${exception.getErrorMessage} }""")
+    }
+
   }
 
-  class SpecialListener(private val logger: org.slf4j.Logger) extends ProgressListener {
+  class LoggingListener(private val logger: org.slf4j.Logger) extends ProgressListener {
     def progressChanged(event: ProgressEvent): Unit = {
       if (event.getEventType == com.amazonaws.event.ProgressEventType.TRANSFER_COMPLETED_EVENT) {
         logger.info("Upload complete.")
